+++ fuzz/fuzz_targets/deserialize_address.rs
@@ -2,7 +2,7 @@ extern crate bitcoin;
 use std::str::FromStr;
 fn do_test(data: &[u8]) {
     let data_str = String::from_utf8_lossy(data);
-    let addr = match bitcoin::util::address::Address::from_str(&data_str) {
+    let addr = match bitcoin::address::Address::from_str(&data_str) {
         Ok(addr) => addr,
         Err(_) => return,
     };
